+++ NEWS.md
@@ -1,4 +1,8 @@
+- Added `chk_all()` function.
 - Added `x_name = NULL` argument to relevant `chk_` functions.
+- Deprecated `chk_files()` and `chk_dirs()` for `chk_file()` and `chk_dir()` 
+which now handle vectors.
+- All error messages except `chk_file()` and `chk_dir()` now start with x_name.
 
 # chk 0.0.1